Put IO functions into helpers module
