+++ docs/src/reference.md
@@ -36,12 +36,6 @@ Modules = [MOLGA.Types]
 Modules = [MOLGA.Configuration]
 ```
 
-## I/O
-
-```@autodocs
-Modules = [MOLGA.IO]
-```
-
 ## Helpers
 
 ```@autodocs
