+++ test/integration/initial_population.jl
@@ -9,7 +9,7 @@ using Random
 
         filename = "temp-testfile-population-$(name).xyz"
 
-        MOLGA.IO.export_xyz(population, filename)
+        MOLGA.Helpers.export_xyz(population, filename)
 
         # check if files are equal
         return success(`diff $filename integration/xyz/population-$name.xyz`)
