+++ test/runtests.jl
@@ -3,7 +3,7 @@ using Test, Documenter
 
 @testset verbose = true "MOLGA.jl" begin
     @testset "Unit Tests" begin
-        include("./unit/io.jl")
+        include("./unit/helpers.jl")
     end
 
     @testset "Integration Tests" begin
