+++ test/unit/helpers.jl
@@ -1,5 +1,5 @@
-@testset verbose = true "IO" begin
-    @test MOLGA.IO.load_xyz("unit/xyz/co2.xyz") == [
+@testset verbose = true "Helpers" begin
+    @test MOLGA.Helpers.load_xyz("unit/xyz/co2.xyz") == [
         BaseAtom(6, [-1.1619373, 0.0, -1.704638554]),
         BaseAtom(8, [-0.610097929, -0.126832008, -0.681728275]),
         BaseAtom(8, [-1.713776672, 0.126832008, -2.727548833]),
@@ -10,7 +10,7 @@
         Atom(1, [-1.674872668, 0.759337000, -0.388923492], 2, 2),
         Atom(1, [-1.674872668, -0.759337000, -0.388923492], 3, 2),
     ]
-    MOLGA.IO.export_xyz(
+    MOLGA.Helpers.export_xyz(
         [
             Structure(; atoms, energy=56.13537, age=10),
             Structure(; atoms, nuclear_repulsion_energy=1.006),
